+++ benchmark/micro_dyadic_multiply.sh
@@ -5,24 +5,29 @@
 
 set -eo pipefail
 
+spath=$(dirname $0)
+. ${spath}/bitstream_dir.sh
+
+########################################
 # FPGA run with individual bitstream
-if [[ -n ${RUN_CHOICE} ]] && [[ ${RUN_CHOICE} -eq 2 ]]
+########################################
+if [[ -z ${RUN_CHOICE} ]] || [[ ${RUN_CHOICE} -eq 2 ]]
 then
-    aocl program acl0 dyadic_multiply.aocx
+    aocl program acl0 ${bitstream_dir}/dyadic_multiply.aocx
 fi
 
 echo ""
-echo "FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 1"
-FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=1 ./bench_dyadic_multiply
+echo "FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 1"
+FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=1 ./bench_dyadic_multiply
 echo ""
-echo "FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 4"
-FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=4 ./bench_dyadic_multiply
+echo "FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 4"
+FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=4 ./bench_dyadic_multiply
 echo ""
-echo "FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 8"
-FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=8 ./bench_dyadic_multiply
+echo "FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 8"
+FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=8 ./bench_dyadic_multiply
 echo ""
-echo "FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 16"
-FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=16 ./bench_dyadic_multiply
+echo "FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 16"
+FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=16 ./bench_dyadic_multiply
 echo ""
-echo "FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 32"
-FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=32 ./bench_dyadic_multiply
+echo "FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY = 32"
+FPGA_BITSTREAM=${bitstream_dir}/dyadic_multiply.aocx FPGA_KERNEL=DYADIC_MULTIPLY BATCH_SIZE_DYADIC_MULTIPLY=32 ./bench_dyadic_multiply