+++ device/lib/hls/mod_ops.h
@@ -19,6 +19,13 @@ ubitwidth_t MultiplyUIntModLazy4(ubitwidth_t x, ubitwidth_t y_operand,
                                  ubitwidth_t modulus);
 ubitwidth_t MultiplyUIntModLazy3(ubitwidth_t x, ubitwidth_t y,
                                  ubitwidth_t modulus);
+ubitwidth_t HLS_MultiplyUInt64Hi(ubitwidth_t x, ubitwidth_t y);
+void HLS_MultiplyUInt52(ubitwidth_t x, ubitwidth_t y, ubitwidth_t* prod_hi,
+                        ubitwidth_t* prod_lo);
+ubitwidth_t HLS_BarrettReduce104(ubitwidth_t input_hi, ubitwidth_t input_lo,
+                                 ubitwidth_t modulus, unsigned long k);
+ubitwidth_t HLS_BarrettReduce128(ubitwidth_t input_hi, ubitwidth_t input_lo,
+                                 ubitwidth_t modulus, unsigned long k);
 #ifdef __cplusplus
 }
 #endif