+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Changes
 
+## Version 1.1
+- Added KeySwitch FPGA kernel support.
+- Added integrated kernel support for DyadicMultiply and KeySwitch.
+- Deprecated standalone NTT and INTT support, which will be removed in the next release.
+- Documentation updates.
+
+Co-Authored-by: @[ymeng-git](https://github.com/ymeng-git) <br/>
+Co-Authored-by: @[yzhou79](https://github.com/yzhou79) <br/>
+Co-Authored-by: @[tgonzalez89-intel](https://github.com/tgonzalez89-intel) <br/>
+
 ## Version 1.0
 - Initial release.
 - Added NTT/INTT/DyadicMultiply FPGA kernel support.